requirements.txt
azure-functions
azure-functions-durable>=1.0.0b6
azure-storage-blob